Refactor examples to utilize `getCachePath` for data directory management and remove redundant `findLatestContractDataFile` function implementations. Add `minimist` as a dependency across multiple examples for improved command-line argument parsing.
