+++ examples/common/package.json
@@ -12,6 +12,7 @@
   "devDependencies": {
     "@types/node": "^22.8.7",
     "copy-file": "^11.0.0",
+    "minimist": "^1.2.8",
     "node-fetch": "^3.3.2",
     "tar": "^7.4.3",
     "typescript": "^5.6.3"
