+++ examples/zeto/src/index.ts
@@ -22,7 +22,7 @@ import erc20Abi from "./zeto-abis/SampleERC20.json";
 import * as fs from 'fs';
 import * as path from 'path';
 import { ContractData } from "./tests/data-persistence";
-import { nodeConnections } from "paladin-example-common";
+import { nodeConnections, getCachePath } from "paladin-example-common";
 
 const logger = console;
 
@@ -279,7 +279,7 @@ async function main(): Promise<boolean> {
   };
 
   // Use command-line argument for data directory if provided, otherwise use default
-  const dataDir = process.argv[2] || path.join(__dirname, '..', 'data');
+  const dataDir = getCachePath();
   if (!fs.existsSync(dataDir)) {
     fs.mkdirSync(dataDir, { recursive: true });
   }
